I found a bug in DTGridView where changes to the underlying data don't become visible when you pop back to a DTGridView from a subsequent navigation controller view.
What actually happens is that the changes are noticed -
addCellWithInfo is called for each cell, but the previous views are
not removed. So if I pop back and forth (making changes to the
underlying data each time) I see the subviews count continually
increase. Since the view is underneath the other views it is not
visible.

Adding a call to removeCellWithInfo for each addCell didn't work
because it seems that some entries in gridCells have lightweight info
objects even if the cell is visible on screen.

The fix to addCellWithInfo feels a little hacky, but given the small number of subviews
onscreen at any one time it will at least be fast.